First 3D scene commit
Added GameWorld.java to hold all the 3D objects. Added Tunnel.java, which will be one of those objects. In application, added createGame() which returns a subscene that will be used as the scene in start. Subscene gets it's root and camera from created GameWorld object. Implemented a testCube method that's executed in the GameWorld constructor. It adds a box to the scene and moves the camera so it's visible.